dev: attend last
Reinstated the function `attend_last`. I had forgotten that in order
for the hooks to work we need a way of knowing who was the one who was
greeted. The only way I can come up with at the moment is to store this
information inside a file and retrieve it when a hook is called.