Adding User
Adding a user is a potential dangerous situation. Someone might do
the promus setup under the administrator's name (your name) and when
using the key they will be registered as you. To avoid this the
user's email and the one pending in the server must match. Otherwise
the whole operation needs to stop.