Mapping ACL
Before checking if a user has access we have to map the acl so that
users are mapped to their emails. Since emails are unique we won't
run into the problem of having two users with the same username.

The acl file will remain the only admin file. The others are
acceptable to be modified by the users in the repository.