DEV: PROMUS_USER
Testing the environmental variable PROMUS_USER. There is a potential
flaw in the way we access the user who is logged in as us. The
current version of promus needs to be changed fast. For this reason
we will move from version 0.1 to 1.0. This may take week but
hopefully once the authorized keys are written correctly we will not
need to provide any more information. There needs to be a way to
access all the users and acls in a clear and straight forward way. We
will branch out for now.