Broken Code & Clone Idea
Some parts of the code is functional now. In the core we have the
modules util, and ssh done up to this point. The git and user modules
need more work. There is an idea that perhaps is worth pursuing. In
the clone command we may want to create a file with pairs of
repositories and clones. In this way when we search for repositories
we will be able to display if we already cloned a repository or not.