Executing Commands
Found a way to execute custom commands. More commands need to be
created (git commands) but this prototype seems to work. The scp
commands work as intended now and only allow to place files in a
certain folder. The send file needs an option which allows the user
to send the file to a shared folder so that other collaborators of
the host can see and get the files.