Command Access
The commands can now be given access privileges. By default everyone
has access but this can be restricted with the use of the file
`~.promus/cmd.py`. We can view the documentation for each command. We
now need to move on to the search command which will allow us to see
which commands are available to the user.