User Git Access
The user git access has been improved. Please see the testing files
for an example of how to write the new acl and how to determine if a
user has access or not. The user module has been deleted and placed
under the promus.core module. More reorganization is needed in the
git module